+++ src/slc_node.cpp
@@ -26,6 +26,7 @@ utilities_impl(unary_op)
 utilities_impl(if_expr)
 utilities_impl(expression)
 utilities_impl(extern_function)
+utilities_impl(lambda)
 utilities_impl(list)
 utilities_impl(literal)
 utilities_impl(formal)
